+++ tests/core/test_model.py
@@ -23,6 +23,7 @@ from unittest.mock import MagicMock
 import pytest
 import pytorch_lightning as pl
 import torch
+from pytorch_lightning import LightningDataModule
 from pytorch_lightning.callbacks import Callback
 from torch import nn, Tensor
 from torch.nn import functional as F
@@ -406,13 +407,32 @@ def test_external_optimizers_torch_optimizer(tmpdir, optim):
         ("cosine_with_hard_restarts_schedule_with_warmup", {"num_warmup_steps": 0.1, "num_cycles": 3}),
     ],
 )
-def test_external_schedulers_provider_hf_transformers(tmpdir, optim, sched):
+@pytest.mark.parametrize("use_datamodule", [False, True])
+@pytest.mark.parametrize("limit", [None, 5, 0.1])
+def test_external_schedulers_provider_hf_transformers(tmpdir, optim, sched, use_datamodule, limit):
     model = nn.Sequential(nn.Flatten(), nn.Linear(28 * 28, 10), nn.LogSoftmax())
     task = ClassificationTask(model, optimizer=deepcopy(optim), lr_scheduler=deepcopy(sched), loss_fn=F.nll_loss)
-    trainer = flash.Trainer(max_epochs=1, limit_train_batches=10, gpus=torch.cuda.device_count())
-    ds = DummyDataset()
-    trainer.fit(task, train_dataloader=DataLoader(ds))
 
+    if limit is not None:
+        batch_count = limit if isinstance(limit, int) else int(limit * 10)
+        trainer = flash.Trainer(max_epochs=1, limit_train_batches=limit)
+    else:
+        batch_count = 10
+        trainer = flash.Trainer(max_epochs=1)
+
+    ds = DummyDataset(num_samples=10)
+
+    if use_datamodule:
+
+        class TestDataModule(LightningDataModule):
+            def train_dataloader(self):
+                return DataLoader(ds)
+
+        trainer.fit(task, datamodule=TestDataModule())
+    else:
+        trainer.fit(task, train_dataloader=DataLoader(ds))
+
+    assert task.get_num_training_steps() == batch_count
     assert isinstance(trainer.optimizers[0], torch.optim.Adadelta)
     assert isinstance(trainer.lr_schedulers[0]["scheduler"], torch.optim.lr_scheduler.LambdaLR)